requirements.txt
Django==1.8.19
GeoNode==2.6.3
MultipartPostHandler==0.1.0
OWSLib==0.16.0
Paver==1.3.4
Pillow==5.1.0
PyYAML==3.12
Shapely==1.5.17
Unidecode==0.04.21
amqp==1.4.9
anyjson==0.3.3
awesome-slugify==1.6.5
beautifulsoup4==4.6.0
billiard==3.3.0.23
boto==2.48.0
celery==3.1.26.post2
certifi==2018.4.16
chardet==3.0.4
configparser==3.5.0
dj-database-url==0.5.0
django-activity-stream==0.6.5
django-appconf==1.0.2
django-autocomplete-light==2.3.3
django-bootstrap-form==3.4
django-bootstrap3-datetimepicker==2.2.3
django-braces==1.13.0
django-celery==3.2.2
django-downloadview==1.9
django-extensions==2.0.7
django-forms-bootstrap==3.1.0
django-friendly-tag-loader==1.2.1
django-geoexplorer==4.0.40
django-guardian==1.4.9
django-haystack==2.8.1
django-js-asset==1.1.0
django-jsonfield==1.0.1
django-jsonfield-compat==0.4.4
django-leaflet==0.24.0
django-modeltranslation==0.12.2
django-mptt==0.9.0
django-nose==1.4.5
django-oauth-toolkit==0.12.0
django-pagination==1.0.7
django-polymorphic==0.9.2
django-storages==1.6.6
django-taggit==0.22.2
django-tastypie==0.13.1
django-treebeard==4.3
elasticsearch==6.2.0
enum34==1.1.6
flake8==3.5.0
geolinks==0.2.0
geonode-agon-ratings==0.3.8
geonode-announcements==1.0.13
geonode-arcrest==10.2
geonode-avatar==2.1.8
geonode-dialogos==1.2
geonode-notification==1.1.3
geonode-user-accounts==1.0.15
geonode-user-messages==0.1.14
gisdata==0.5.4
gsconfig==1.0.9
gsimporter==1.0.0
httplib2==0.11.3
idna==2.6
kombu==3.0.37
lxml==3.6.2
mccabe==0.6.1
nose==1.3.7
numpy==1.14.4
oauthlib==2.0.1
pep8==1.7.1
pinax-theme-bootstrap==8.0.1
pinax-theme-bootstrap-account==1.0b2
psycopg2==2.7.4
pycodestyle==2.3.1
pycsw==2.2.0
pyelasticsearch==1.4.1
pyflakes==1.6.0
pygdal==2.1.0.3
pyproj==1.9.5.1
python-dateutil==2.7.3
python-mimeparse==1.6.0
pytz==2018.4
regex==2018.6.9
requests==2.18.4
simplejson==3.15.0
six==1.10.0
slugify==0.0.1
transifex-client==0.12.5
typing==3.6.4
urllib3==1.22
xmltodict==0.10.2
